DRY-up *PendingTransactionSorter tests
Based on previous work of @garyschulte
